requirements.txt
Flask==0.12.2
flask_sqlalchemy==2.3.2
marshmallow-sqlalchemy==0.13.2
flask-marshmallow==0.8.0
flask-migrate==2.1.1
flask-security==3.0.0
Flask-BabelEx==0.9.3
Flask-Babel==0.11.2
flask-bcrypt==0.7.1
flask-moment==0.6.0
Flask-JSGlue==0.3.1
celery==4.1.0
flask_restful==0.3.6
pyopenssl==17.5.0
requests==2.18.4
ipython==5.5.0
marshmallow==2.15.0
flask-cors==3.0.3
pyrabbit==1.1.0
flower==0.9.2
CMRESHandler==1.0.0
simplejson==3.13.2
PyJWT==1.6.0
psycopg2==2.7.4
minio==3.0.3
python-dotenv==0.8.2
gitpython==2.1.9
#xhtml2pdf==0.2.2
#dotenv==0.0.5
# Install eventlet if you are working on Windows
#eventlet
